+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.x branch
 ## 2.30 branch
 ### 2.30.1
+* if Recaptcha is not used, an exception is no longer throwned when configuration
+    keys are missing;
 * added `CreateSamplePostCommand`, this creates a sample post during installation;
 * fixed the `MenuBuilderHelper` and the admin sidebar, menus are now better displayed.